+++ src/context/index.tsx
@@ -6,6 +6,10 @@ import {
   useState,
 } from "react";
 import { getProvider } from "src/providers/factory";
+import {
+  refreshAccessToken,
+  revokeAccessToken,
+} from "src/providers/identidad/authorization";
 import { IUser } from "src/types/user";
 import { IAuthContext } from "./types";
 
@@ -36,31 +40,81 @@ function AuthProvider(props: AuthProviderProps) {
   const [isLoading, setIsLoading] = useState(true);
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   const [user, setUser] = useState<IUser>();
+  const [accessToken, setAccessToken] = useState<string>();
 
-  useEffect(() => {
+  const loadUserFromStorage = () => {
     const savedUser = localStorage.getItem("user");
 
     if (savedUser) {
       setUser(JSON.parse(savedUser));
       setIsAuthenticated(true);
     }
+  };
+
+  const refreshTokens = async () => {
+    const savedAccessToken = localStorage.getItem("accessToken");
+    const refreshToken = localStorage.getItem("refreshToken");
+
+    if (savedAccessToken && realm && clientSecret && refreshToken) {
+      const refreshTokenResponse = await refreshAccessToken(
+        savedAccessToken,
+        realm,
+        clientId,
+        clientSecret,
+        refreshToken
+      );
+
+      if (!refreshTokenResponse) return;
+
+      localStorage.setItem("accessToken", refreshTokenResponse.accessToken);
+      localStorage.setItem("refreshToken", refreshTokenResponse.refreshToken);
+      localStorage.setItem("expiresIn", refreshTokenResponse.expiresIn);
+
+      setAccessToken(refreshTokenResponse.accessToken);
+    }
+  };
+
+  const setupRefreshInterval = () => {
+    const interval = setInterval(
+      refreshTokens,
+      Number(localStorage.getItem("expiresIn"))
+    );
+    return () => clearInterval(interval);
+  };
+
+  useEffect(() => {
+    loadUserFromStorage();
+    refreshTokens();
     setIsLoading(false);
+
+    return setupRefreshInterval();
   }, []);
 
   const loginWithRedirect = useCallback(async () => {
     const selectedProvider = getProvider(provider);
 
-    const user = await selectedProvider.loginWithRedirect({
+    const sessionData = await selectedProvider.loginWithRedirect({
       clientId,
       clientSecret,
       realm,
       authorizationParams,
     });
 
-    if (!user) return;
+    if (!sessionData) return;
 
-    setUser(user);
-    localStorage.setItem("user", JSON.stringify(user));
+    setUser(sessionData.user);
+    setAccessToken(sessionData.accessToken);
+
+    localStorage.setItem("user", JSON.stringify(sessionData.user));
+    localStorage.setItem(
+      "accessToken",
+      JSON.stringify(sessionData.accessToken)
+    );
+
+    if (sessionData.refreshToken) {
+      localStorage.setItem("refreshToken", sessionData.refreshToken);
+      localStorage.setItem("expiresIn", JSON.stringify(sessionData.expiresIn));
+    }
 
     setIsAuthenticated(true);
     setIsLoading(false);
@@ -68,22 +122,29 @@ function AuthProvider(props: AuthProviderProps) {
     window && window.location.replace(authorizationParams.redirectUri);
   }, [user, authorizationParams.redirectUri]);
 
-  const logout = useCallback(() => {
+  const logout = useCallback(async () => {
+    if (accessToken && realm) {
+      revokeAccessToken(accessToken, realm);
+    }
+
     localStorage.removeItem("user");
+    localStorage.removeItem("accessToken");
     setUser(undefined);
+    setAccessToken(undefined);
     setIsAuthenticated(false);
-  }, []);
+  }, [accessToken]);
 
   const authContext = useMemo(
     () => ({
       user,
+      accessToken,
       isLoading,
       isAuthenticated,
 
       loginWithRedirect,
       logout,
     }),
-    [user, isLoading, isAuthenticated, loginWithRedirect, logout]
+    [user, accessToken, isLoading, isAuthenticated, loginWithRedirect, logout]
   );
 
   return (