+++ src/context/types.ts
@@ -2,6 +2,7 @@ import { IUser } from "src/types/user";
 
 interface IAuthContext {
   user?: IUser;
+  accessToken?: string;
   isLoading: boolean;
   isAuthenticated: boolean;